+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## CHANGELOG
 
-### UNRELEASED
+### 0.18.0 - 2020-07-23
 
  - Add support for redis geospatial data type